refactor: Centralize Heart Rate and Workout Type Definitions
Centralized redundant Heart Rate and Workout Type Definitions to eliminate 'Type Slop'.

- Moved local interfaces into centralized domain files.
- Merged `lib/hrm/zones.ts` (HrZone), `utils/hr-zones.ts` (UserHrZones), and `types/index.ts` (HrTileProps) into `types/heart-rate.ts`.
- Synchronized `WorkoutItem` definitions between `types/index.ts` and the persistence layer types in `types/workout.ts`.
- This ensures a single source of truth for domain interfaces.
