+++ hooks/useWorkoutSessionManager.ts
@@ -5,8 +5,8 @@ import {
   workoutSessionStorage,
   WorkoutSessionData,
   HrDataPoint,
-  HrZoneName,
 } from '../lib/workout-session-storage'
+import { HrZoneName } from '../lib/shared/hr-zones'
 import { v4 as uuidv4 } from 'uuid'
 import { calculateHrZone } from '../lib/hrm/zones'
 import { estimateMaxHr } from '../lib/hrm/utils'
